+++ src/test/java/uk/gov/di/ipv/cri/common/library/persistence/SessionItemTest.java
@@ -6,6 +6,7 @@ import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
+import uk.gov.di.ipv.cri.common.library.persistence.item.ContextClaim;
 import uk.gov.di.ipv.cri.common.library.persistence.item.EvidenceRequest;
 import uk.gov.di.ipv.cri.common.library.persistence.item.SessionItem;
 import uk.gov.di.ipv.cri.common.library.service.SessionService;
@@ -18,8 +19,6 @@ import static org.mockito.Mockito.when;
 
 @ExtendWith(MockitoExtension.class)
 class SessionItemTest {
-    private static final String INT_USER_CONTEXT = "international_user";
-
     @Mock private DataStore<SessionItem> dataStore;
 
     @InjectMocks private SessionService sessionService;
@@ -40,7 +39,7 @@ class SessionItemTest {
         evidenceRequest.setIdentityFraudScore(6);
         sessionItem.setEvidenceRequest(evidenceRequest);
 
-        sessionItem.setContext(INT_USER_CONTEXT);
+        sessionItem.setContext(ContextClaim.INTERNATIONAL_USER);
 
         when(dataStore.getItem(any())).thenReturn(sessionItem);
     }
@@ -60,6 +59,6 @@ class SessionItemTest {
         assertEquals(5, evidenceRequest.getActivityHistoryScore());
         assertEquals(6, evidenceRequest.getIdentityFraudScore());
 
-        assertEquals(INT_USER_CONTEXT, sessionItem.getContext());
+        assertEquals(ContextClaim.INTERNATIONAL_USER, sessionItem.getContext());
     }
 }